llm/9_transcript/requirements.txt
openai
python-dotenv
streamlit
moviepy
pydub